+++ src/hooks/useEtherspotBalances.ts
@@ -1,4 +1,4 @@
-import { AccountBalance } from '@etherspot/prime-sdk';
+import { AccountBalance } from '@etherspot/prime-sdk/dist/sdk/data';
 import { useMemo } from 'react';
 
 // hooks